feat(logging): add last run communication log file
OpenTrack is a giant pice of SHIT. The console is pretty much useless,
it slows everything down significantly (WTF?), it's lost unless you
manually copy it right before OpenTrack terminates, the find button
does nothing etc.

This does the same but allows you to use your favorite text editor to
see the communication that just happend, it persists if it crashed, it
doesn't slow the simulation down etc.