fix(outputs)!: issues with wrongly low delays
The file format OpenTrack exports is fucked up. 0.0 is both zero delay
and no data (WTF), -1.0 is both 1 second too early and never reached
(WTF again). It's necessary to investigate other values to decide what
these numbers mean.

The CSV were moved from the outputs dir to it's parent directory with
the outpust basename in it's filename. This makes it much easier to
locate the files in Google Drive.